CB-77 Implement Order Table

+++ src/main/java/com/ooadprojectserver/restaurantmanagement/dto/request/order/OrderRequest.java
@@ -10,5 +10,5 @@ import lombok.*;
 @NoArgsConstructor
 public class OrderRequest {
     @JsonProperty("table_id")
-    private Long tableId;
+    private Integer tableId;
 }
